refactor: extract TransientToastAction.swift from TransientToastTypes.swift
Splits the `TransientToastAction` struct (button descriptor) out of
TransientToastTypes.swift so that file holds only the style enum.

Byte-identical move of lines 17-35. Both files import Foundation.

TransientToastTypes.swift: 35 → 16 lines. Tests: 667.
